requirements.txt
django
bash
dj_database_url
requests
python-dotenv
psycopg2-binary
djangorestframework-simplejwt